requirements.txt
pyinaturalist>=0.18.0.dev4
arcgis>=2.0.1
func_timeout==4.3.5
pytz~=2022.6
python-dateutil~=2.8.2

git+https://github.com/behave/behave@v1.2.7.dev2
behave_html_formatter>=0.9.10
pyhamcrest>=2.0.4
factory_boy>=3.2.1

flake8==6.0.0
